+++ epoxy-paging3/src/main/java/com/airbnb/epoxy/paging3/PagedListModelCache.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.airbnb.epoxy.paging
+package com.airbnb.epoxy.paging3
 
 import android.annotation.SuppressLint
 import android.os.Handler
